+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v0.1.3
 
+* New fsio_copy_file functions.
 * New fsio_path_exists function.
 * New fsio_file_exists function.
 * Free all memory in tests